+++ agent/main.go
@@ -33,7 +33,7 @@ import (
 	"time"
 )
 
-var Version = "2.2.86"
+var Version = "2.2.87"
 
 const selfUpgradeLockTimeout = 10 * time.Minute
 const iperf3IdleTimeout = 3 * time.Minute